Add aion rewards calculator compound table

+++ aion_pool/src/Miningcore/Blockchain/Aion/Configuration/AionPoolPaymentExtraConfig.cs
@@ -49,7 +49,7 @@ namespace Miningcore.Blockchain.Aion.Configuration
         public string PrivateKey { get; set; }
         public bool SendTransactionsUsingPrivateKey { get; set; }
         public bool EnableMinerMinimumPayment { get; set; }
-        public decimal NetworkForkBlock { get; set; }
+        public long NetworkForkBlock { get; set; }
         public decimal InitialSupply { get; set; }
     }
 }
